devtools/client/shared/components/reps/test/mochitest/test_reps_event.html

<!-- This Source Code Form is subject to the terms of the Mozilla Public
   - License, v. 2.0. If a copy of the MPL was not distributed with this
   - file, You can obtain one at http://mozilla.org/MPL/2.0/. -->
<!DOCTYPE HTML>
<html>
<!--
Test Event rep
-->
<head>
  <meta charset="utf-8">
  <title>Rep test - Event</title>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css">
</head>
<body>
<pre id="test">
<script src="head.js" type="application/javascript"></script>
<script type="application/javascript">
window.onload = Task.async(function* () {
  const {
    REPS,
    MODE,
    getRep,
    getSelectableInInspectorGrips,
  } = browserRequire("devtools/client/shared/components/reps/reps");
  let { Event } = REPS;

  try {
    // Test that correct rep is chosen
    is(getRep(getGripStub("testEvent")), Event.rep, "Rep correctly selects Event Rep");

    yield testEvent();
    yield testMouseEvent();
    yield testKeyboardEvent();
    yield testKeyboardEventWithModifiers();
    yield testMessageEvent();

    yield testOnDomNodeMouseOver();
    yield testOnDomNodeMouseOut();
    yield testOnDomNodeInspectIconClick();

    yield testObjectLink();
  } catch (e) {
    ok(false, "Got an error: " + DevToolsUtils.safeErrorString(e));
  } finally {
    SimpleTest.finish();
  }

  function testEvent() {
    const renderedComponent = renderComponent(Event.rep, {
      object: getGripStub("testEvent"),
    });
    is(renderedComponent.textContent,
      "beforeprint { target: Window, isTrusted: true, " +
      "currentTarget: Window, 8 more… }",
      "Event rep has expected text content for an event");
  }

  function testMouseEvent() {
    const renderedComponent = renderComponent(Event.rep, {
      object: getGripStub("testMouseEvent")
    });
    is(renderedComponent.textContent,
      "click { target: div#test, clientX: 62, clientY: 18, 3 more… }",
      "Event rep has expected text content for a mouse event");

    const longRenderedComponent = renderComponent(Event.rep, {
      object: getGripStub("testMouseEvent"),
      mode: MODE.LONG,
    });
    is(longRenderedComponent.textContent,
      "click { target: div#test, buttons: 0, clientX: 62, clientY: 18, layerX: 0, " +
      "layerY: 0 }",
      "Event rep has expected text content for a mouse event in long mode");
  }

  function testKeyboardEvent() {
    const renderedComponent = renderComponent(Event.rep, {
      object: getGripStub("testKeyboardEvent")
    });
    is(renderedComponent.textContent,
       "keyup { target: body, key: \"Control\", charCode: 0, 1 more… }",
       "Event rep has expected text content for a keyboard event");

    const longRenderedComponent = renderComponent(Event.rep, {
      object: getGripStub("testKeyboardEvent"),
      mode: MODE.LONG,
    });
    is(longRenderedComponent.textContent,
       `keyup { target: body, key: "Control", charCode: 0, keyCode: 17 }`,
       "Event rep has expected text content for a keyboard event in long mode");
  }

  function testKeyboardEventWithModifiers() {
    const renderedComponent = renderComponent(Event.rep, {
      object: getGripStub("testKeyboardEventWithModifiers"),
      mode: MODE.LONG,
    });

    is(renderedComponent.textContent,
       `keyup Meta-Shift { target: body, key: "M", charCode: 0, keyCode: 77 }`,
       "Event rep has expected text content for a keyboard event with modifiers " +
       "in long mode");
  }

  function testMessageEvent() {
    const renderedComponent = renderComponent(Event.rep, {
      object: getGripStub("testMessageEvent")
    });
    is(renderedComponent.textContent,
       `message { target: Window, isTrusted: false, data: "test data", ` +
       "8 more… }",
       "Event rep has expected text content for a message event");

    const longRenderedComponent = renderComponent(Event.rep, {
      object: getGripStub("testMessageEvent"),
      mode: MODE.LONG,
    });
    is(longRenderedComponent.textContent,
       `message { target: Window, isTrusted: false, data: "test data", ` +
       `origin: "null", lastEventId: "", source: Window, ports: Array, ` +
       `currentTarget: Window, eventPhase: 2, bubbles: false, 1 more… }`,
       "Event rep has expected text content for a message event in long mode");
  }

  function testOnDomNodeMouseOver() {
    const stub = getGripStub("testMouseEvent");
    const grips = getSelectableInInspectorGrips(stub);

    is(grips.length, 1, "the stub has one node grip");

    let mouseOverValue;
    let onDOMNodeMouseOver = (object) => {
      mouseOverValue = object;
    };
    const renderedComponent = renderComponent(Event.rep, {
      object: stub,
      onDOMNodeMouseOver,
    });

    const node = renderedComponent.querySelector(".objectBox-node");
    TestUtils.Simulate.mouseOver(node);

    is(mouseOverValue, grips[0], "onDOMNodeMouseOver is called with " +
      "the expected argument when mouseover is fired on the Rep");
  }

  function testOnDomNodeMouseOut() {
    const stub = getGripStub("testMouseEvent");
    const grips = getSelectableInInspectorGrips(stub);
    is(grips.length, 1, "the stub has one node grip");

    let called = false;
    let onDOMNodeMouseOut = (object) => {
      called = true;
    };
    const renderedComponent = renderComponent(Event.rep, {
      object: stub,
      onDOMNodeMouseOut,
    });

    const node = renderedComponent.querySelector(".objectBox-node");
    TestUtils.Simulate.mouseOut(node);

    is(called, true, "onDOMNodeMouseOut is called when mouseout is fired on the Rep");
  }

  function testOnDomNodeInspectIconClick() {
    const stub = getGripStub("testMouseEvent");
    const grips = getSelectableInInspectorGrips(stub);
    is(grips.length, 1, "the stub has one node grip");

    let inspectIconClickedValue = null;
    let onInspectIconClick = (object) => {
      inspectIconClickedValue = object;
    };

    const renderedComponent = renderComponent(Event.rep, {
      object: stub,
      onInspectIconClick,
    });

    const icon = renderedComponent.querySelector(".open-inspector");
    ok(icon !== null,
      "There is an inspect icon when the node is connected to the DOM tree");

    TestUtils.Simulate.click(icon);

    is(inspectIconClickedValue, grips[0],
      "onInspectIconClick is called with the expected argument " +
      "when the inspect icon is clicked");
  }

  function testObjectLink() {
    const renderedComponent = renderComponent(Event.rep, {
      object: getGripStub("testMouseEvent"),
      objectLink: (props, ...children) => React.DOM.span({},
        "*", ...children, "*"),
    });
    is(renderedComponent.textContent,
      "*click** { *target: div#test, clientX: 62, clientY: 18, *3 more…** }*",
      "Event rep has expected text content when an objectLink is passed as a prop");

    const longRenderedComponent = renderComponent(Event.rep, {
      object: getGripStub("testMouseEvent"),
      mode: MODE.LONG,
      objectLink: (props, ...children) => React.DOM.span({},
        "*", ...children, "*"),
    });
    is(longRenderedComponent.textContent,
      "*click** { *target: div#test, buttons: 0, clientX: 62, clientY: 18, layerX: 0, " +
      "layerY: 0* }*",
      "Event rep has expected text content in long mode when an objectLink is passed " +
      "as a prop");
  }

  function getGripStub(name) {
    switch (name) {
      case "testEvent":
        return {
          "type": "object",
          "class": "Event",
          "actor": "server1.conn23.obj35",
          "extensible": true,
          "frozen": false,
          "sealed": false,
          "ownPropertyLength": 1,
          "preview": {
            "kind": "DOMEvent",
            "type": "beforeprint",
            "properties": {
              "isTrusted": true,
              "currentTarget": {
                "type": "object",
                "class": "Window",
                "actor": "server1.conn23.obj37",
                "extensible": true,
                "frozen": false,
                "sealed": false,
                "ownPropertyLength": 760,
                "preview": {
                  "kind": "ObjectWithURL",
                  "url": "http://example.com"
                }
              },
              "eventPhase": 2,
              "bubbles": false,
              "cancelable": false,
              "defaultPrevented": false,
              "timeStamp": 1466780008434005,
              "originalTarget": {
                "type": "object",
                "class": "Window",
                "actor": "server1.conn23.obj38",
                "extensible": true,
                "frozen": false,
                "sealed": false,
                "ownPropertyLength": 760,
                "preview": {
                  "kind": "ObjectWithURL",
                  "url": "http://example.com"
                }
              },
              "explicitOriginalTarget": {
                "type": "object",
                "class": "Window",
                "actor": "server1.conn23.obj39",
                "extensible": true,
                "frozen": false,
                "sealed": false,
                "ownPropertyLength": 760,
                "preview": {
                  "kind": "ObjectWithURL",
                  "url": "http://example.com"
                }
              },
              "NONE": 0
            },
            "target": {
              "type": "object",
              "class": "Window",
              "actor": "server1.conn23.obj36",
              "extensible": true,
              "frozen": false,
              "sealed": false,
              "ownPropertyLength": 760,
              "preview": {
                "kind": "ObjectWithURL",
                "url": "http://example.com"
              }
            }
          }
        };

      case "testMouseEvent":
        return {
          "type": "object",
          "class": "MouseEvent",
          "actor": "server1.conn20.obj39",
          "extensible": true,
          "frozen": false,
          "sealed": false,
          "ownPropertyLength": 1,
          "preview": {
            "kind": "DOMEvent",
            "type": "click",
            "properties": {
              "buttons": 0,
              "clientX": 62,
              "clientY": 18,
              "layerX": 0,
              "layerY": 0
            },
            "target": {
              "type": "object",
              "class": "HTMLDivElement",
              "actor": "server1.conn20.obj40",
              "extensible": true,
              "frozen": false,
              "sealed": false,
              "ownPropertyLength": 0,
              "preview": {
                "kind": "DOMNode",
                "nodeType": 1,
                "nodeName": "div",
                "isConnected": true,
                "attributes": {
                  "id": "test"
                },
                "attributesLength": 1
              }
            }
          }
        };

      case "testKeyboardEvent":
        return {
          "type": "object",
          "class": "KeyboardEvent",
          "actor": "server1.conn21.obj49",
          "extensible": true,
          "frozen": false,
          "sealed": false,
          "ownPropertyLength": 1,
          "preview": {
            "kind": "DOMEvent",
            "type": "keyup",
            "properties": {
              "key": "Control",
              "charCode": 0,
              "keyCode": 17
            },
            "target": {
              "type": "object",
              "class": "HTMLBodyElement",
              "actor": "server1.conn21.obj50",
              "extensible": true,
              "frozen": false,
              "sealed": false,
              "ownPropertyLength": 0,
              "preview": {
                "kind": "DOMNode",
                "nodeType": 1,
                "nodeName": "body",
                "attributes": {},
                "attributesLength": 0
              }
            },
            "eventKind": "key",
            "modifiers": []
          }
        };

      case "testKeyboardEventWithModifiers":
        return {
          "type": "object",
          "class": "KeyboardEvent",
          "actor": "server1.conn21.obj49",
          "extensible": true,
          "frozen": false,
          "sealed": false,
          "ownPropertyLength": 1,
          "preview": {
            "kind": "DOMEvent",
            "type": "keyup",
            "properties": {
              "key": "M",
              "charCode": 0,
              "keyCode": 77
            },
            "target": {
              "type": "object",
              "class": "HTMLBodyElement",
              "actor": "server1.conn21.obj50",
              "extensible": true,
              "frozen": false,
              "sealed": false,
              "ownPropertyLength": 0,
              "preview": {
                "kind": "DOMNode",
                "nodeType": 1,
                "nodeName": "body",
                "attributes": {},
                "attributesLength": 0
              }
            },
            "eventKind": "key",
            "modifiers": ["Meta", "Shift"]
          }
        };

      case "testMessageEvent":
        return {
          "type": "object",
          "class": "MessageEvent",
          "actor": "server1.conn3.obj34",
          "extensible": true,
          "frozen": false,
          "sealed": false,
          "ownPropertyLength": 1,
          "preview": {
            "kind": "DOMEvent",
            "type": "message",
            "properties": {
              "isTrusted": false,
              "data": "test data",
              "origin": "null",
              "lastEventId": "",
              "source": {
                "type": "object",
                "class": "Window",
                "actor": "server1.conn3.obj36",
                "extensible": true,
                "frozen": false,
                "sealed": false,
                "ownPropertyLength": 760,
                "preview": {
                  "kind": "ObjectWithURL",
                  "url": ""
                }
              },
              "ports": {
                "type": "object",
                "class": "Array",
                "actor": "server1.conn3.obj37",
                "extensible": true,
                "frozen": false,
                "sealed": false,
                "ownPropertyLength": 0
              },
              "currentTarget": {
                "type": "object",
                "class": "Window",
                "actor": "server1.conn3.obj38",
                "extensible": true,
                "frozen": false,
                "sealed": false,
                "ownPropertyLength": 760,
                "preview": {
                  "kind": "ObjectWithURL",
                  "url": ""
                }
              },
              "eventPhase": 2,
              "bubbles": false,
              "cancelable": false
            },
            "target": {
              "type": "object",
              "class": "Window",
              "actor": "server1.conn3.obj35",
              "extensible": true,
              "frozen": false,
              "sealed": false,
              "ownPropertyLength": 760,
              "preview": {
                "kind": "ObjectWithURL",
                "url": "http://example.com"
              }
            }
          }
        };

    }
  }
});
</script>
</pre>
</body>
</html>
